TEST LOGIN AND SIGNUP
This is only meant for testing and will not be used in the final build of our project. This was source code was online and will only be used for data integration with the backend while the rest of our website is being worked on.
Original Website Source: https://medium.com/technoetics/create-basic-login-forms-using-create-react-app-module-in-reactjs-511b9790dede

Also, download Node_Module.zip from here:
https://drive.google.com/file/d/1suvR-aQnzosVYPHtChYlM0tN7pbpMBAI/view?usp=sharing
Extract it into the frontend folder.
if not uploaded

For setting up:
cd ~/frontend
run: npm install --save material-ui axios react-tap-event-plugin
Do: npm rebuild
Then: npm start